EuroProject/requirements.txt
kagglehub==1.0.0
rank-bm25==0.2.2
nltk==3.9.1
pandas==2.2.2
sentence-transformers==5.3.0
numpy==2.0.2
deep-translator==1.11.4
faiss-cpu